requirements.txt
Flask==2.2.3
selenium==4.9.1
webdriver_manager==3.8.5
pywin32==306